+++ src/config.rs
@@ -1,0 +1,16 @@
+use std::fs;
+use std::path::Path;
+
+use serde::Deserialize;
+
+use crate::macros::ok_or_return;
+
+#[derive(Debug, Default, Deserialize)]
+pub struct Config {}
+
+impl Config {
+    pub fn load(path: &Path) -> Self {
+        let content = ok_or_return!(fs::read_to_string(path), Self::default());
+        ok_or_return!(toml::from_str(&content), Self::default())
+    }
+}