+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Procedure when bumping the version number:
 - `ephemeral` config option
 - `data_dir` config option
 - `euph.rooms.<name>.username` config option
+- `euph.rooms.<name>.force_username` config option
 - `euph.rooms.<name>.password` config option
 
 ## v0.3.0 - 2022-08-22