+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Procedure when bumping the version number:
 - `euph.rooms.<name>.password` config option
 - Key binding to change rooms sort order
 - Key bindings to connect to/disconnect from all rooms
-- Key bindings to move to parent/top-level parent
+- Key bindings to move to parent/root
 
 ### Changed
 - Some key bindings in the rooms list