+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Procedure when bumping the version number:
 - Key binding to change rooms sort order
 - Key bindings to connect to/disconnect from all rooms
 - Key bindings to move to parent/root message
+- Key bindings to view and open links in a message
 
 ### Changed
 - Some key bindings in the rooms list